requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
pyttsx3==2.90
fpdf2==2.7.8 